requirements.txt
playwright>=1.37.0
beautifulsoup4>=4.12.2
markdownify>=0.11.6
pyyaml>=6.0
lxml>=4.9.3
certifi==2025.1.31
charset-normalizer==3.4.1
greenlet==3.1.1
idna==3.10
pyee==12.1.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
typing_extensions==4.13.2
urllib3==2.4.0 